+++ lib/frontmatter-linter.mjs
@@ -1,9 +1,9 @@
-const frontmatter = require('@github-docs/frontmatter');
-const semver = require('semver');
+import frontmatter from '@github-docs/frontmatter';
+import semver from 'semver';
 
 const releasedOrLaterStages = ['released', 'recommended'];
 
-module.exports = class FrontmatterLinter {
+export default class FrontmatterLinter {
   constructor(stages = [], teams = []) {
     this.rules = {
       stage: {
