+++ lib/frontmatter-updater.mjs
@@ -1,7 +1,7 @@
-const frontmatter = require('@github-docs/frontmatter');
-const merge = require('deepmerge');
+import frontmatter from '@github-docs/frontmatter';
+import merge from 'deepmerge';
 
-module.exports = class FrontmatterUpdater {
+export default class FrontmatterUpdater {
   constructor(markdown) {
     this.markdown = markdown;
   }
