+++ lib/result-reporter.mjs
@@ -1,10 +1,10 @@
-const chalk = require('chalk');
+import chalk from 'chalk';
 
 function log() {
   console.log(...arguments);
 }
 
-module.exports = class ResultReporter {
+export default class ResultReporter {
   constructor(results) {
     this.results = results;
   }
