+++ package.json
@@ -5,7 +5,7 @@
   "private": true,
   "scripts": {
     "lint": "eslint .",
-    "test": "mocha test/**/*.js && mocha test/**/*.mjs"
+    "test": "mocha test/**/*.mjs"
   },
   "author": "Katie Gengler",
   "license": "MIT",
