+++ scripts/rfc-frontmatter.mjs
@@ -1,4 +1,9 @@
-const argv = require('yargs').command(
+import yargs from 'yargs';
+import { hideBin } from 'yargs/helpers'
+import { frontmatter } from '../lib/frontmatter.mjs';
+import { readFileSync } from 'node:fs';
+
+const argv = yargs(hideBin(process.argv)).command(
   '* path',
   'return the frontmatter of the RFC as JSON',
   (yargs) => {
@@ -11,8 +16,6 @@ const argv = require('yargs').command(
   }
 ).argv;
 
-const frontmatter = require('../lib/frontmatter.js');
-const { readFileSync } = require('fs');
 
 let markdown = readFileSync(argv.path, 'utf8');
 const { data, errors } = frontmatter(markdown);
@@ -20,7 +23,6 @@ const { data, errors } = frontmatter(markdown);
 if (errors.length) {
   console.error(JSON.stringify(errors));
   process.exitCode = 1;
-  return;
+} else {
+  console.log(JSON.stringify(data));
 }
-
-console.log(JSON.stringify(data));
